requirements.txt
fastapi
uvicorn[standard]
playwright
textblob
openai
python-dotenv
pytest